compiler/pyproject.toml
[project]
name = "compiler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "description-generator",
    "pydantic>=2.11.3",
    "pygithub>=2.6.1",
    "pytz>=2025.2",
]

[tool.uv.sources]
description-generator = { git = "https://github.com/Stephen-Hallett/Repo-Description-Generator.git" }
